+++ pyproject.toml
@@ -98,12 +98,13 @@ types-toml = "^0.10.8"
 types-typed-ast = "^1.5.8"
 types-ujson = "^5.6.0"
 types-protobuf = "^4.23.0.1"
-mypy = "1.9.0"
+mypy = {git = "https://github.com/python/mypy.git", rev = "master"}
 pytest-mypy-plugins = "3.0.0"
 poetry-plugin-export = "^1.6.0"
 # another bug in poetry
 urllib3 = "<2"
 graphlib_backport = {version = "*", python = "<3.9", optional = false}
+inline-snapshot = "^0.10.1"
 
 [tool.poetry.group.integrations]
 optional = true
