--- tests/pyright/test_federation_params.py
@@ -1,36 +1,0 @@
-from .utils import Result, requires_pyright, run_pyright, skip_on_windows
-
-pytestmark = [skip_on_windows, requires_pyright]
-
-
-CODE = """
-import strawberry
-
-
-@strawberry.federation.type(name="User")
-class UserModel:
-    name: str
-
-
-UserModel(name="Patrick")
-UserModel(n="Patrick")
-"""
-
-
-def test_pyright():
-    results = run_pyright(CODE)
-
-    assert results == [
-        Result(
-            type="error",
-            message='Argument missing for parameter "name"',
-            line=11,
-            column=1,
-        ),
-        Result(
-            type="error",
-            message='No parameter named "n"',
-            line=11,
-            column=11,
-        ),
-    ]
