+++ tests/type_checkers/test_fields.py
@@ -1,0 +1,73 @@
+from inline_snapshot import snapshot
+
+from .utils.marks import requires_mypy, requires_pyright, skip_on_windows
+from .utils.typecheck import Result, typecheck
+
+pytestmark = [skip_on_windows, requires_pyright, requires_mypy]
+
+
+CODE = """
+import strawberry
+
+
+@strawberry.type
+class User:
+    name: str
+
+
+User(name="Patrick")
+User(n="Patrick")
+
+reveal_type(User)
+reveal_type(User.__init__)
+"""
+
+
+def test():
+    results = typecheck(CODE)
+
+    assert results.pyright == snapshot(
+        [
+            Result(
+                type="error",
+                message='Argument missing for parameter "name"',
+                line=11,
+                column=1,
+            ),
+            Result(type="error", message='No parameter named "n"', line=11, column=6),
+            Result(
+                type="information",
+                message='Type of "User" is "type[User]"',
+                line=13,
+                column=13,
+            ),
+            Result(
+                type="information",
+                message='Type of "User.__init__" is "(self: User, *, name: str) -> None"',
+                line=14,
+                column=13,
+            ),
+        ]
+    )
+    assert results.mypy == snapshot(
+        [
+            Result(
+                type="error",
+                message='Unexpected keyword argument "n" for "User"',
+                line=11,
+                column=1,
+            ),
+            Result(
+                type="note",
+                message='Revealed type is "def (*, name: builtins.str) -> mypy_test.User"',
+                line=13,
+                column=13,
+            ),
+            Result(
+                type="note",
+                message='Revealed type is "def (self: mypy_test.User, *, name: builtins.str)"',
+                line=14,
+                column=13,
+            ),
+        ]
+    )
