+++ tests/type_checkers/test_fields_resolver.py
@@ -1,0 +1,77 @@
+from inline_snapshot import snapshot
+
+from .utils.marks import requires_mypy, requires_pyright, skip_on_windows
+from .utils.typecheck import Result, typecheck
+
+pytestmark = [skip_on_windows, requires_pyright, requires_mypy]
+
+
+CODE = """
+import strawberry
+
+def get_user_age() -> int:
+    return 0
+
+
+@strawberry.type
+class User:
+    name: str
+    age: int = strawberry.field(resolver=get_user_age)
+
+
+User(name="Patrick")
+User(n="Patrick")
+
+reveal_type(User)
+reveal_type(User.__init__)
+"""
+
+
+def test():
+    results = typecheck(CODE)
+
+    assert results.pyright == snapshot(
+        [
+            Result(
+                type="error",
+                message='Argument missing for parameter "name"',
+                line=15,
+                column=1,
+            ),
+            Result(type="error", message='No parameter named "n"', line=15, column=6),
+            Result(
+                type="information",
+                message='Type of "User" is "type[User]"',
+                line=17,
+                column=13,
+            ),
+            Result(
+                type="information",
+                message='Type of "User.__init__" is "(self: User, *, name: str) -> None"',
+                line=18,
+                column=13,
+            ),
+        ]
+    )
+    assert results.mypy == snapshot(
+        [
+            Result(
+                type="error",
+                message='Unexpected keyword argument "n" for "User"',
+                line=15,
+                column=1,
+            ),
+            Result(
+                type="note",
+                message='Revealed type is "def (*, name: builtins.str) -> mypy_test.User"',
+                line=17,
+                column=13,
+            ),
+            Result(
+                type="note",
+                message='Revealed type is "def (self: mypy_test.User, *, name: builtins.str)"',
+                line=18,
+                column=13,
+            ),
+        ]
+    )
