+++ tests/type_checkers/utils/result.py
@@ -1,0 +1,16 @@
+from dataclasses import dataclass
+from typing import Literal
+
+ResultType = Literal[
+    "error",
+    "information",
+    "note",
+]
+
+
+@dataclass
+class Result:
+    type: ResultType
+    message: str
+    line: int
+    column: int
